Revert "Merge branch 'dev' into 201014_JM_Port_Child_Row_Rendering"
This reverts commit 77da07978ef39cb5753f901e521d64f365ae480e, reversing
changes made to 6e886d5638d5cf76c1d7fea6924caf56e2198449.

+++ src/MudBlazor/Components/Popover/MudPopoverBase.cs
@@ -2,15 +2,17 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;
 
 namespace MudBlazor;
 
-
+#nullable enable
 /// <summary>
-/// Base class for popover components such as <see cref="MudPopover"/>, handling creation and lifecycle through the <see cref="IPopoverService"/>.
+/// Base class for implementing Popover component.
 /// </summary>
 /// <remarks>
 /// This class provides a base implementation for a Popover component. It implements the <see cref="IPopover"/> interface
@@ -18,10 +20,17 @@ namespace MudBlazor;
 /// </remarks>
 public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposable
 {
+    [Obsolete("For Legacy compatibility mode only, will be removed in v7.")]
+    private MudPopoverHandler? _handler;
+
     private bool _afterFirstRender;
 
     /// <inheritdoc />
-    public virtual Guid Id { get; } = Guid.NewGuid();
+    public virtual Guid Id { get; [Obsolete("Set is only needed for legacy mode only. Remove in v7.")] private set; } = Guid.NewGuid();
+
+    [Inject]
+    [Obsolete($"Replaced by {nameof(PopoverService)}. Will be removed in v7.")]
+    public IMudPopoverService Service { get; set; } = null!;
 
     [Inject]
     protected IPopoverService PopoverService { get; set; } = null!;
@@ -49,7 +58,18 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
     /// <inheritdoc />
     protected override async Task OnInitializedAsync()
     {
-        await PopoverService.CreatePopoverAsync(this);
+        if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
+#pragma warning disable CS0618 // Type or member is obsolete
+        {
+            _handler = Service.Register(ChildContent ?? new RenderFragment((x) => { }));
+            _handler.SetComponentBaseParameters(this, PopoverClass, PopoverStyles, Open);
+            Id = _handler.Id;
+        }
+#pragma warning restore CS0618 // Type or member is obsolete
+        else
+        {
+            await PopoverService.CreatePopoverAsync(this);
+        }
 
         await base.OnInitializedAsync();
     }
@@ -61,7 +81,19 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
 
         if (_afterFirstRender)
         {
-            await PopoverService.UpdatePopoverAsync(this);
+            if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
+#pragma warning disable CS0618 // Type or member is obsolete
+            {
+                if (_handler is not null)
+                {
+                    await _handler.UpdateFragmentAsync(ChildContent, this, PopoverClass, PopoverStyles, Open);
+                }
+            }
+#pragma warning restore CS0618 // Type or member is obsolete
+            else
+            {
+                await PopoverService.UpdatePopoverAsync(this);
+            }
         }
     }
 
@@ -70,7 +102,21 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
     {
         if (firstRender)
         {
-            await PopoverService.UpdatePopoverAsync(this);
+            if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
+#pragma warning disable CS0618 // Type or member is obsolete
+            {
+                if (_handler is not null)
+                {
+                    await _handler.Initialize();
+                    await Service.InitializeIfNeeded();
+                    await _handler.UpdateFragmentAsync(ChildContent, this, PopoverClass, PopoverStyles, Open);
+                }
+            }
+#pragma warning restore CS0618 // Type or member is obsolete
+            else
+            {
+                await PopoverService.UpdatePopoverAsync(this);
+            }
 
             _afterFirstRender = true;
         }
@@ -86,7 +132,16 @@ public abstract class MudPopoverBase : MudComponentBase, IPopover, IAsyncDisposa
         {
             if (IsJSRuntimeAvailable)
             {
-                await PopoverService.DestroyPopoverAsync(this);
+#pragma warning disable CS0618 // Type or member is obsolete
+                if (PopoverService.PopoverOptions.Mode == PopoverMode.Legacy)
+                {
+                    await Service.Unregister(_handler);
+                }
+#pragma warning restore CS0618 // Type or member is obsolete
+                else
+                {
+                    await PopoverService.DestroyPopoverAsync(this);
+                }
             }
         }
         catch (JSDisconnectedException) { }
